requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.11
requests==2.27.1
beautifulsoup4==4.11.1